Refine zone update handling in refactored world services
